MT#62181 use std::string for string building
Automatic memory management and less error prone as it avoids manual
pointer manipulations.

No functional change.

+++ core/sip/sip_parser.cpp
@@ -43,7 +43,7 @@
 #include <memory>
 using std::unique_ptr;
 
-sip_msg::sip_msg(const char* msg_buf, int msg_len)
+sip_msg::sip_msg(const char* msg_buf, size_t msg_len)
     : hdrs(),
       to(NULL),
       from(NULL),
